+++ src/pages/Home.jsx
@@ -3,11 +3,24 @@ import ModalComponent from '../components/feed/questionFeed/ModalComponent';
 import profile from '../assets/images/samples/profile-sample.png';
 import { ThemeProvider } from 'styled-components';
 import theme from '../styles/@shared/theme';
+import { useState } from 'react';
 
 function Home() {
+  const [modalIsOpen, setModalIsOpen] = useState(false);
+
+  const showModal = () => {
+    setModalIsOpen(true);
+  };
+  const closeModal = () => {
+    setModalIsOpen(false);
+  };
+
   return (
     <ThemeProvider theme={theme}>
-      <ModalComponent profileImg={profile} name="go5rae" />
+      <button onClick={showModal}>열기</button>
+      {modalIsOpen && (
+        <ModalComponent isOpen={showModal} onRequestClose={closeModal} profileImg={profile} name="go5rae" />
+      )}
     </ThemeProvider>
   );
 }
